Added Pickaxe to Inventory

+++ OctoAwesome/OctoAwesome.Client/Components/HudComponent.cs
@@ -103,7 +103,7 @@ namespace OctoAwesome.Client.Components
         {
             get
             {
-                return new Index2((int)GraphicsDevice.Viewport.Width, (int)GraphicsDevice.Viewport.Height);
+                return new Index2(GraphicsDevice.Viewport.Width, GraphicsDevice.Viewport.Height);
             }
         }
